+++ app/src/main/kotlin/ac/mdiq/podcini/ui/screens/EpisodeInfoScreen.kt
@@ -28,7 +28,7 @@ import ac.mdiq.podcini.ui.actions.PlayLocalActionButton
 import ac.mdiq.podcini.ui.actions.StreamActionButton
 import ac.mdiq.podcini.ui.activity.MainActivity
 import ac.mdiq.podcini.ui.activity.MainActivity.Companion.mainNavController
-import ac.mdiq.podcini.ui.activity.MainActivity.Screens
+import ac.mdiq.podcini.ui.screens.Screens
 import ac.mdiq.podcini.ui.compose.ChaptersDialog
 import ac.mdiq.podcini.ui.compose.ChooseRatingDialog
 import ac.mdiq.podcini.ui.compose.CustomTextStyles